Add missing changes in the last merge.

+++ com.unity.render-pipelines.universal/Shaders/2D/Light2D-Shape-Volumetric.shader
@@ -34,23 +34,23 @@ Shader "Hidden/Light2D-Shape-Volumetric"
             struct Varyings
             {
                 float4  positionCS  : SV_POSITION;
-                float4  color       : COLOR;
-                float2  uv          : TEXCOORD0;
+                half4   color       : COLOR;
+                half2   uv          : TEXCOORD0;
 
                 SHADOW_COORDS(TEXCOORD1)
             };
 
-            float4 _LightColor;
-            float  _FalloffDistance;
-            float4 _FalloffOffset;
-            float  _VolumeOpacity;
-            float  _InverseHDREmulationScale;
+            half4 _LightColor;
+            half  _FalloffDistance;
+            half4 _FalloffOffset;
+            half  _VolumeOpacity;
+            half  _InverseHDREmulationScale;
 
 #ifdef SPRITE_LIGHT
             TEXTURE2D(_CookieTex);			// This can either be a sprite texture uv or a falloff texture
             SAMPLER(sampler_CookieTex);
 #else
-            uniform float  _FalloffIntensity;
+            uniform half  _FalloffIntensity;
             TEXTURE2D(_FalloffLookup);
             SAMPLER(sampler_FalloffLookup);
 #endif
